+++ R/RLearner_classif_xgboost.R
@@ -120,7 +120,9 @@ predictLearner.classif.xgboost = function(.learner, .model, .newdata, ...) {
     }
   } else { #multiclass
     if (.learner$par.vals$objective  == "multi:softmax") {
-      return(factor(p, levels = cls)) #special handling for multi:softmax which directly predicts class levels
+      p = as.factor(p) #special handling for multi:softmax which directly predicts class levels
+      levels(p) = cls
+      return(p)
     } else {
       p = matrix(p, nrow = length(p) / nc, ncol = nc, byrow = TRUE)
       colnames(p) = cls
@@ -137,7 +139,7 @@ predictLearner.classif.xgboost = function(.learner, .model, .newdata, ...) {
 
 #' @export
 getFeatureImportanceLearner.classif.xgboost = function(.learner, .model, ...) {
-  mod = getLearnerModel(.model)
+  mod = getLearnerModel(.model, more.unwrap = TRUE)
   imp = xgboost::xgb.importance(feature_names = .model$features,
     model = mod, ...)